Student Management System
This is my first ever project created in 1st year of engineering. We created this project in C language. The data was stored in a file using file handling in C. This project gives the functionality like adding a student data in database, removing a student's record from database, searching for a specific student's record and also reading the data of all student. We can store the academic details of a student as well as his information. We created structure having all these values and then just updated it as the information comes. While adding the information we opened the file in write mode and updated that file as well.